Fix build for Intel@userpc60.
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@22019 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/common/DistributedTetrahedralMesh.cpp
@@ -1661,8 +1661,7 @@ unsigned DistributedTetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::GetNearestNodeIndex
 
     // Return the index of the closest node to the current point
     // In the distributed case, we'll have to do an AllReduce
-    double global_best_node_index = this->mNodes[best_node_index]->GetIndex();
-
+    unsigned global_best_node_index = this->mNodes[best_node_index]->GetIndex();
 
     // This is a handy data structure that will work with MPI_DOUBLE_INT data type.
     // todo #2507 Is this the best place for this struct?
